VFS: moving permission checks from VFS to file servers
Having VFS perform access permission checks forces file servers to work with
the Unix permission model. This is not a problem for classic file systems
like ext2 or ISO 9660, but distributed file systems (CODA for instance) use
different permission models with remote credentials.

This patch moves the burden of permission checking from VFS to file servers
(with a default Unix permission bits checking available in libfsdriver).

Note: this is a work-in-progress. Although the POSIX test suite is passed,
this is nowhere near production-ready (the biggest issues for now are inode
leaks and almost no real testing done yet).