PXE boot generation script
Leveraging the new image generation framework, MINIX has now turn-key
PXE booting capability. The booting process uses PXELINUX, licensed
under GPL2+.

Due to the lack of NFS support, a ramdisk is provided as a root device.
The generated ramdisk is quite huge for a network boot, but it can be
trimmed down at a later date.